requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.2
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jinja2==3.1.6
markupsafe==3.0.3
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-decouple==3.8
requests==2.32.5
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
